+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added support for parallel database loading.
 - Added support for MS SQL and ClickHouse keywords.
 
 ### Changed